BASE OBS: Move base observation handling into its own file
